+++ src/main/java/kitchenpos/products/tobe/domain/repository/ProductRepository.java
@@ -1,0 +1,9 @@
+package kitchenpos.products.tobe.domain.repository;
+
+import kitchenpos.products.tobe.domain.model.Product;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ProductRepository extends JpaRepository<Product, Long> {
+}
