+++ src/main/resources/application.properties
@@ -3,3 +3,9 @@ spring.datasource.data=classpath*:db/data.sql
 spring.datasource.schema=classpath*:db/schema.sql
 spring.h2.console.enabled=true
 spring.http.encoding.force=true
+
+## Hibernate
+spring.jpa.show-sql=true
+spring.jpa.properties.hibernate.format_sql=true
+
+spring.jpa.hibernate.ddl-auto=validate
